+++ docs/pages/documentation/data/Table.vue
@@ -11,21 +11,28 @@
                     <b-switch v-model="isBordered">Bordered</b-switch>
                     <b-switch v-model="isStriped">Striped</b-switch>
                     <b-switch v-model="isNarrowed">Narrowed</b-switch>
-                    <b-switch v-model="isSelectable">Selectable</b-switch>
                     <b-switch v-model="isCheckable">Checkable</b-switch>
                     <b-switch v-model="isPaginated">Paginated</b-switch>
                     <b-switch v-model="isPaginationSimple">Simple pagination</b-switch>
                     <b-switch v-model="hasMobileCards">Rows as cards on mobile</b-switch>
                 </div>
                 <div class="block">
-                    <b-field label="Items per page">
-                        <b-select v-model="perPage">
-                            <b-option value="5">5</b-option>
-                            <b-option value="10">10</b-option>
-                            <b-option value="15">15</b-option>
-                            <b-option value="20">20</b-option>
-                        </b-select>
-                    </b-field>
+                    <b-select v-model="perPage" style="display: inline-block">
+                        <option value="5">5 per page</option>
+                        <option value="10">10 per page</option>
+                        <option value="15">15 per page</option>
+                        <option value="20">20 per page</option>
+                    </b-select>
+                    <button class="button" @click="selected = {}"
+                        :disabled="Object.keys(selected).length === 0">
+                        <b-icon icon="clear"></b-icon>
+                        <span>Clear Selected</span>
+                    </button>
+                    <button class="button" @click="checkedRows = []"
+                        :disabled="checkedRows.length === 0">
+                        <b-icon icon="clear"></b-icon>
+                        <span>Clear Checked</span>
+                    </button>
                 </div>
 
                 <div class="example">
@@ -34,119 +41,117 @@
                         :bordered="isBordered"
                         :striped="isStriped"
                         :narrowed="isNarrowed"
-                        :selectable="isSelectable"
                         :checkable="isCheckable"
                         :mobile-cards="hasMobileCards"
                         :paginated="isPaginated"
                         :per-page="perPage"
                         :pagination-simple="isPaginationSimple"
-                        default-sort="first_name"
-                        render-html
-                        @check="checked"
-                        @select="selected">
+                        default-sort="user.first_name"
+                        :selected.sync="selected"
+                        :checked-rows.sync="checkedRows">
 
-                        <b-table-column field="id" label="ID" width="40" sortable numeric></b-table-column>
-                        <b-table-column field="first_name" label="First Name" sortable></b-table-column>
-                        <b-table-column field="last_name" label="Last Name" sortable></b-table-column>
-                        <b-table-column field="date" label="Date" sortable :format="formatDate"></b-table-column>
-                        <b-table-column field="gender" label="Gender" component="gender-icon"></b-table-column>
+                        <template scope="props">
+                            <b-table-column field="id" label="ID" width="40" sortable numeric>
+                                {{ props.row.id }}
+                            </b-table-column>
+
+                            <b-table-column field="user.first_name" label="First Name" sortable>
+                                {{ props.row.user.first_name }}
+                            </b-table-column>
+
+                            <b-table-column field="user.last_name" label="Last Name" sortable>
+                                {{ props.row.user.last_name }}
+                            </b-table-column>
+
+                            <b-table-column field="date" label="Date" sortable
+                                v-html="formatDate(props.row.date)">
+                            </b-table-column>
+
+                            <b-table-column field="gender" label="Gender">
+                                <b-icon
+                                    pack="fa"
+                                    :icon="props.row.gender === 'Male' ? 'mars' : 'venus'">
+                                </b-icon>
+                                {{ props.row.gender }}
+                            </b-table-column>
+                        </template>
                     </b-table>
                 </div>
             </div>
             <div class="column">
-                <h3 class="subtitle">Selected item</h3>
+                <h3 class="subtitle">Selected</h3>
                 <pre class="content">
-{{ selItem }}
+{{ selected }}
                 </pre>
 
-                <h3 class="subtitle">Checked items</h3>
+                <h3 class="subtitle">Checked rows</h3>
                 <pre>
-{{ checkItems }}
+{{ checkedRows }}
                 </pre>
             </div>
         </div>
         <pre class="content" v-highlight><code class="html">{{ template | pre }}</code></pre>
         <pre class="content" v-highlight><code class="javascript">{{ code | pre }}</code></pre>
-        <h2 class="subtitle">GenderIcon.vue</h2>
-        <pre v-highlight><code class="vue">{{ componentCode | pre }}</code></pre>
 
         <hr>
 
         <h2 class="subtitle">Table properties</h2>
-        <b-table
-            :data="tableProps"
-            default-sort="name"
-            render-html>
-            <b-table-column
-                field="name"
-                label="Name">
-            </b-table-column>
-            <b-table-column
-                field="description"
-                label="Description"
-                width="620">
-            </b-table-column>
-            <b-table-column
-                field="type"
-                label="Type">
-            </b-table-column>
-            <b-table-column
-                field="values"
-                label="Values">
-            </b-table-column>
-            <b-table-column
-                field="default"
-                label="Default">
-            </b-table-column>
+        <b-table :data="tableProps" default-sort="name">
+            <template scope="props">
+                <b-table-column field="name" label="Name"
+                    v-html="props.row.name">
+                </b-table-column>
+                <b-table-column field="description" label="Description" width="620"
+                    v-html="props.row.description">
+                </b-table-column>
+                <b-table-column field="type" label="Type"
+                    v-html="props.row.type">
+                </b-table-column>
+                <b-table-column field="values" label="Values"
+                    v-html="props.row.values">
+                </b-table-column>
+                <b-table-column field="default" label="Default"
+                    v-html="props.row.default">
+                </b-table-column>
+            </template>
         </b-table>
 
         <h2 class="subtitle">Table events</h2>
-        <b-table
-            :data="tableEvents"
-            default-sort="name"
-            render-html>
-            <b-table-column
-                field="name"
-                label="Name">
-            </b-table-column>
-            <b-table-column
-                field="description"
-                label="Description"
-                width="620">
-            </b-table-column>
-            <b-table-column
-                field="parameters"
-                label="Parameters">
-            </b-table-column>
+        <b-table :data="tableEvents" default-sort="name">
+            <template scope="props">
+                <b-table-column field="name" label="Name"
+                    v-html="props.row.name">
+                </b-table-column>
+                <b-table-column field="description" label="Description" width="620"
+                    v-html="props.row.description">
+                </b-table-column>
+                <b-table-column field="parameters" label="Parameters"
+                    v-html="props.row.parameters">
+                </b-table-column>
+            </template>
         </b-table>
 
         <hr>
-        <h2 class="subtitle">Table Column properties</h2>
-        <b-table
-            :data="columnProps"
-            default-sort="name"
-            render-html>
-            <b-table-column
-                field="name"
-                label="Name">
-            </b-table-column>
-            <b-table-column
-                field="description"
-                label="Description"
-                width="620">
-            </b-table-column>
-            <b-table-column
-                field="type"
-                label="Type">
-            </b-table-column>
-            <b-table-column
-                field="values"
-                label="Values">
-            </b-table-column>
-            <b-table-column
-                field="default"
-                label="Default">
-            </b-table-column>
+
+        <h2 class="subtitle">Column properties</h2>
+        <b-table :data="columnProps" default-sort="name">
+            <template scope="props">
+                <b-table-column field="name" label="Name"
+                    v-html="props.row.name">
+                </b-table-column>
+                <b-table-column field="description" label="Description" width="620"
+                    v-html="props.row.description">
+                </b-table-column>
+                <b-table-column field="type" label="Type"
+                    v-html="props.row.type">
+                </b-table-column>
+                <b-table-column field="values" label="Values"
+                    v-html="props.row.values">
+                </b-table-column>
+                <b-table-column field="default" label="Default"
+                    v-html="props.row.default">
+                </b-table-column>
+            </template>
         </b-table>
 
     </div>
@@ -155,20 +160,15 @@
 <script>
     import tableData from '../../../assets/data_test.json'
 
-    import Vue from 'vue'
-    import GenderIcon from '../../../components/GenderIcon.vue'
-    Vue.component('GenderIcon', GenderIcon)
-
     export default {
         data() {
             return {
                 tableData,
-                checkItems: [],
-                selItem: {},
+                checkedRows: [],
+                selected: {},
                 isBordered: false,
                 isStriped: false,
                 isNarrowed: false,
-                isSelectable: false,
                 isCheckable: false,
                 hasMobileCards: true,
                 isPaginated: true,
@@ -204,11 +204,11 @@
                         default: '<code>false</code>'
                     },
                     {
-                        name: '<code>selectable</code>',
-                        description: 'Clicks on rows will select (single)',
-                        type: 'Boolean',
+                        name: '<code>selected</code>',
+                        description: 'Set which row is selected, use the <code>.sync</code> modifier to make it two-way binding',
+                        type: 'Object',
                         values: '—',
-                        default: '<code>false</code>'
+                        default: '—'
                     },
                     {
                         name: '<code>checkable</code>',
@@ -218,18 +218,18 @@
                         default: '<code>false</code>'
                     },
                     {
+                        name: '<code>checked-rows</code>',
+                        description: 'Set which rows are checked, use the <code>.sync</code> modifier to make it two-way binding',
+                        type: 'Array',
+                        values: '—',
+                        default: '—'
+                    },
+                    {
                         name: '<code>mobile-cards</code>',
                         description: 'Rows now appears as cards on mobile',
                         type: 'Boolean',
                         values: '—',
                         default: '<code>true</code>'
-                    },
-                    {
-                        name: '<code>render-html</code>',
-                        description: 'Cells now renders HTML, <b>do not use it with user-generated content</b>, read more about <a href="https://vuejs.org/v2/guide/syntax.html#Raw-HTML" target="_blank">here</a>',
-                        type: 'Boolean',
-                        values: '—',
-                        default: '<code>false</code>'
                     },
                     {
                         name: '<code>backend-sorting</code>',
@@ -254,7 +254,7 @@
                     },
                     {
                         name: '<code>per-page</code>',
-                        description: 'How many items per page (if <code>paginated</code>)',
+                        description: 'How many rows per page (if <code>paginated</code>)',
                         type: 'Number',
                         values: '—',
                         default: '<code>20</code>'
@@ -300,7 +300,14 @@
                 columnProps: [
                     {
                         name: '<code>label</code>',
-                        description: 'Column header text',
+                        description: '<b>Required</b>, column header text',
+                        type: 'String',
+                        values: '—',
+                        default: '—'
+                    },
+                    {
+                        name: '<code>field</code>',
+                        description: 'Key of the object the column is attributed, used for sorting',
                         type: 'String',
                         values: '—',
                         default: '—'
@@ -309,13 +316,6 @@
                         name: '<code>width</code>',
                         description: 'Column fixed width in pixels',
                         type: 'Number',
-                        values: '—',
-                        default: '—'
-                    },
-                    {
-                        name: '<code>field</code>',
-                        description: 'Key of the object the column is attributed',
-                        type: 'String',
                         values: '—',
                         default: '—'
                     },
@@ -339,21 +339,6 @@
                         type: 'Function (a: Object, b: Object)',
                         values: '—',
                         default: '—'
-                    },
-                    {
-                        name: '<code>format</code>',
-                        description: 'Format cell content',
-                        type: 'Function (value: Any, row: Object)',
-                        values: '—',
-                        default: '—'
-                    },
-                    {
-                        name: '<code>component</code>',
-                        description: `Injects a Vue component in the cell, useful to format cells with <b>user-generated content</b>.
-                            The component will receive a <code>data</code> prop with the cell value and a <code>row</code> prop with the row object`,
-                        type: 'String',
-                        values: '—',
-                        default: '—'
                     }
                 ],
                 template: `
@@ -362,29 +347,42 @@
                         :bordered="isBordered"
                         :striped="isStriped"
                         :narrowed="isNarrowed"
-                        :selectable="isSelectable"
                         :checkable="isCheckable"
                         :mobile-cards="hasMobileCards"
                         :paginated="isPaginated"
                         :per-page="perPage"
                         :pagination-simple="isPaginationSimple"
-                        default-sort="first_name"
-                        render-html
-                        @check="checked"
-                        @select="selected">
+                        default-sort="user.first_name"
+                        :selected.sync="selected"
+                        :checked-rows.sync="checkedRows">
 
-                        <b-table-column field="id" label="ID" width="40" sortable numeric></b-table-column>
-                        <b-table-column field="first_name" label="First Name" sortable></b-table-column>
-                        <b-table-column field="last_name" label="Last Name" sortable></b-table-column>
-                        <b-table-column field="date" label="Date" sortable :format="formatDate"></b-table-column>
-                        <b-table-column field="gender" label="Gender" component="gender-icon"></b-table-column>
+                        <template scope="props">
+                            <b-table-column field="id" label="ID" width="40" sortable numeric>
+                                {{ props.row.id }}
+                            </b-table-column>
 
+                            <b-table-column field="user.first_name" label="First Name" sortable>
+                                {{ props.row.user.first_name }}
+                            </b-table-column>
+
+                            <b-table-column field="user.last_name" label="Last Name" sortable>
+                                {{ props.row.user.last_name }}
+                            </b-table-column>
+
+                            <b-table-column field="date" label="Date" sortable
+                                v-html="formatDate(props.row.date)">
+                            </b-table-column>
+
+                            <b-table-column field="gender" label="Gender">
+                                <b-icon
+                                    pack="fa"
+                                    :icon="props.row.gender === 'Male' ? 'mars' : 'venus'">
+                                </b-icon>
+                                {{ props.row.gender }}
+                            </b-table-column>
+                        </template>
                     </b-table>`,
                 code: `
-                    import Vue from 'vue'
-                    import GenderIcon from 'components/GenderIcon.vue' // This is a custom component
-                    Vue.component('GenderIcon', GenderIcon)
-
                     export default {
                         data() {
                             return {
@@ -396,12 +394,11 @@
                                     {"id":5,"first_name":"Anne","last_name":"Lee","date":"2016-12-06 14:38:38","gender":"Female"},
                                     ...
                                 ],
-                                checkItems: [],
-                                selItem: {},
+                                checkedRows: [],
+                                selected: {},
                                 isBordered: false,
                                 isStriped: false,
                                 isNarrowed: false,
-                                isSelectable: false,
                                 isCheckable: false,
                                 hasMobileCards: true,
                                 isPaginated: true,
@@ -410,55 +407,18 @@
                             }
                         },
                         methods: {
-                            checked(items, item) {
-                                this.checkItems = items
-                            },
-                            selected(item) {
-                                this.selItem = item
-                            },
                             formatDate(value, row) {
-                                return \`<span class="tag is-primary">
+                                return \`<span class="tag is-success">
                                     \${new Date(value).toLocaleDateString()}
                                 </span>\`
                             }
                         }
-                    }`,
-                componentCode: `
-                    <template>
-                        <span>
-                            <b-icon :icon="icon" pack="fa"></b-icon>
-                            {{ data }}
-                        </span>
-                    </template>
-
-                    <script>
-                        export default {
-                            props: {
-                                data: String
-                            },
-                            computed: {
-                                icon() {
-                                    switch (this.data) {
-                                        case 'Male': return 'mars'
-                                        case 'female': return 'venus'
-                                        default: return 'question'
-                                    }
-                                }
-                            }
-                        }
-                    <\/script>
-                `
+                    }`
             }
         },
         methods: {
-            checked(items, item) {
-                this.checkItems = items
-            },
-            selected(item) {
-                this.selItem = item
-            },
             formatDate(value, row) {
-                return `<span class="tag is-primary">
+                return `<span class="tag is-success">
                     ${new Date(value).toLocaleDateString()}
                 </span>`
             }